+++ examples/service/verify/access_token/create/access_token_create_example.go
@@ -1,0 +1,41 @@
+package main
+
+import (
+	"log"
+	"os"
+
+	"github.com/RJPearson94/twilio-sdk-go"
+	v2 "github.com/RJPearson94/twilio-sdk-go/service/verify/v2"
+	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/access_tokens"
+	"github.com/RJPearson94/twilio-sdk-go/session/credentials"
+)
+
+var verifySession *v2.Verify
+
+func init() {
+	creds, err := credentials.New(credentials.Account{
+		Sid:       os.Getenv("TWILIO_ACCOUNT_SID"),
+		AuthToken: os.Getenv("TWILIO_AUTH_TOKEN"),
+	})
+	if err != nil {
+		log.Panicf("%s", err.Error())
+	}
+
+	verifySession = twilio.NewWithCredentials(creds).Verify.V2
+}
+
+func main() {
+	_, err := verifySession.
+		Service("VAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX").
+		AccessTokens.
+		Create(&access_tokens.CreateAccessTokenInput{
+			Identity:   "Test User",
+			FactorType: "push",
+		})
+
+	if err != nil {
+		log.Panicf("%s", err.Error())
+	}
+
+	log.Println("Access Token created")
+}