+++ service/verify/v2/service/api_op_client.go
@@ -3,6 +3,7 @@ package service
 
 import (
 	"github.com/RJPearson94/twilio-sdk-go/client"
+	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/access_tokens"
 	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/rate_limit"
 	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/rate_limits"
 	"github.com/RJPearson94/twilio-sdk-go/service/verify/v2/service/verification"
@@ -17,6 +18,7 @@ type Client struct {
 
 	sid string
 
+	AccessTokens      *access_tokens.Client
 	RateLimit         func(string) *rate_limit.Client
 	RateLimits        *rate_limits.Client
 	Verification      func(string) *verification.Client
@@ -36,6 +38,9 @@ func New(client *client.Client, properties ClientProperties) *Client {
 
 		sid: properties.Sid,
 
+		AccessTokens: access_tokens.New(client, access_tokens.ClientProperties{
+			ServiceSid: properties.Sid,
+		}),
 		RateLimit: func(rateLimitSid string) *rate_limit.Client {
 			return rate_limit.New(client, rate_limit.ClientProperties{
 				ServiceSid: properties.Sid,