+++ CHANGELOG.md
@@ -20,6 +20,7 @@ FEATURES
 - **New Resource:** Service Rate Limit Bucket
 - **New Resource:** Verification
 - **New Resource:** Verification Check
+- **New Resource:** Access Tokens
 
 ## v0.5.0 (2020-10-16)